Function achat::chat_with_announce::handle_connection

pub async fn handle_connection<Reader, Writer>(
    addr: SocketAddr, 
    reader: Reader, 
    writer: Writer, 
    tx: Sender<(String, SocketAddr)>, 
    rx: Receiver<(String, SocketAddr)>, 
    topic_rx: Receiver<String>
) -> Result<()> where
    Reader: AsyncRead + Unpin,
    Writer: AsyncWrite + Unpin, 

Monitor the reader, rx and topic_rx for messages. When receiving bytes on reader, forward them on the tokio::sync::broadcast::Sender. When receiving a message on rx, where the source socket address is not our own, forward it on writer (else, discard it). When the content of topic_rx changed, fetch it, then format and forward it on writer.

Termination

If EOF is signalled on reader by Ok(0), terminate the future. If an error or None is encountered, terminate the future.