Function achat::echo::handle_connection_manually

pub async fn handle_connection_manually<Reader, Writer>(
    reader: Reader, 
    writer: Writer
) -> Result<()> where
    Reader: AsyncRead + Unpin,
    Writer: AsyncWrite + Unpin, 

Manually loop, recv and send on reader and writer.

Termination

In case the reader has no more bytes (read_line returned Ok(0)), terminate the future.

Errors

Returns an error if something goes wrong while reading or writing.